bems/dashboard-bms/be-3/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.3 
PyJWT==2.7.0
Flask-Cors==5.0.0
mysql-connector-python
Werkzeug==2.2.3
mysqlclient

